Added MKL diagonalization with FFTW *and* CULA diagonalization with FFTW implementations
Now all of the files in the following directories produce the same
results as the Fortran code:

    mkl/fftw/
    cula/fftw/
    cula/cufft/

Would be nice to get a fully functioning code in /cula/nofft/ so that
I can compare all 4 implementations. It would also be nice to move the
entire function v_of_rho() to the GPU, to see if that could be sped
up. Alternatively, may consider storing FFT data for all 4 bands up
front, rather than storing 4 different times.